Moved subclasses to general file
