--- BlaBlaMessengerServer/src/data_structures/Conference/ConferenceId.java
@@ -1,10 +1,0 @@
-package data_structures.Conference;
-
-import data_structures.DataObject;
-import java.util.UUID;
-
-public class ConferenceId extends DataObject {
-    public UUID Id;
-    
-    public ConferenceId() { Id = UUID.randomUUID(); }
-}
