--- BlaBlaMessengerServer/src/data_structures/Conference/ConferenceName.java
@@ -1,9 +1,0 @@
-package data_structures.Conference;
-
-import data_structures.DataObject;
-
-public class ConferenceName extends DataObject {
-    public String Name;
-    
-    public ConferenceName( String name ) { Name = name; }
-}
