--- BlaBlaMessengerServer/src/data_structures/Contact/ContactId.java
@@ -1,10 +1,0 @@
-package data_structures.Contact;
-
-import data_structures.DataObject;
-import java.util.UUID;
-
-public class ContactId extends DataObject {
-    public UUID Id;
-
-    public ContactId() { Id = UUID.randomUUID(); }
-}
