--- BlaBlaMessengerServer/src/data_structures/Contact/ContactName.java
@@ -1,9 +1,0 @@
-package data_structures.Contact;
-
-import data_structures.DataObject;
-
-public class ContactName extends DataObject {
-    public String Name;
-    
-    public ContactName( final String name  ) { Name = name; }
-}
