+++ BlaBlaMessengerServer/src/data_structures/Contacts.java
@@ -1,6 +1,5 @@
 package data_structures;
 
-import data_structures.Contact.Contact;
 import java.util.ArrayList;
 
 public class Contacts extends DataObject {
