--- BlaBlaMessengerServer/src/data_structures/File/FileData.java
@@ -1,9 +1,0 @@
-package data_structures.File;
-
-import data_structures.DataObject;
-
-public class FileData extends DataObject {
-    public byte[] Data;
-    
-    public FileData( final byte[] data ) { Data = data; }
-}
