--- BlaBlaMessengerServer/src/data_structures/File/FileName.java
@@ -1,9 +1,0 @@
-package data_structures.File;
-
-import data_structures.DataObject;
-
-public class FileName extends DataObject {
-    public String Name;
-    
-    public FileName( String name ) { Name = name; }
-}
