--- BlaBlaMessengerServer/src/data_structures/FileIdNamePair/FileId.java
@@ -1,10 +1,0 @@
-package data_structures.FileIdNamePair;
-
-import data_structures.DataObject;
-import java.util.UUID;
-
-public class FileId extends DataObject {
-    public UUID Id;
-    
-    public FileId() { Id = UUID.randomUUID(); }
-}
