+++ BlaBlaMessengerServer/src/data_structures/FileIdNamePairs.java
@@ -1,6 +1,5 @@
 package data_structures;
 
-import data_structures.FileIdNamePair.FileIdNamePair;
 import java.util.ArrayList;
 
 public class FileIdNamePairs extends DataObject {
